@@ -8,13 +8,13 @@ repositories {
 }
 
 android {
-    compileSdkVersion 24
-    buildToolsVersion "24.0.1"
+    compileSdkVersion 29
+    buildToolsVersion '29.0.3'
 
     defaultConfig {
         applicationId "com.google.firebase.udacity.friendlychat"
-        minSdkVersion 16
-        targetSdkVersion 24
+        minSdkVersion 29
+        targetSdkVersion 29
         versionCode 1
         versionName "1.0"
     }
@@ -35,9 +35,9 @@ dependencies {
     compile fileTree(dir: 'libs', include: ['*.jar'])
     testCompile 'junit:junit:4.12'
 
-    compile 'com.android.support:design:24.2.0'
-    compile 'com.android.support:appcompat-v7:24.2.0'
-
+    compile 'com.google.android.material:material:1.0.0'
+    implementation 'androidx.appcompat:appcompat:1.0.0'
+    implementation 'com.google.firebase:firebase-database:19.3.0'
     // Displaying images
     compile 'com.github.bumptech.glide:glide:3.6.1'
 }
